refactor(admin): PhotoEditTab 데스크톱 컴포넌트를 components/desktop으로 이동
