+++ frontend/src/pages/AdminPage/tabs/PhotoEditTab/PhotoEditTab.tsx
@@ -7,7 +7,7 @@ import useMixpanelTrack from '@/hooks/Mixpanel/useMixpanelTrack';
 import useTrackPageView from '@/hooks/Mixpanel/useTrackPageView';
 import { ContentSection } from '@/pages/AdminPage/components/ContentSection/ContentSection';
 import { ClubDetail } from '@/types/club';
-import { FeedImageGrid } from './components/FeedImageGrid/FeedImageGrid';
+import { FeedImageGrid } from './components/desktop/FeedImageGrid/FeedImageGrid';
 import { useDragSort } from './hooks/useDragSort';
 import { useFeedItems } from './hooks/useFeedItems';
 import * as Styled from './PhotoEditTab.styles';
