+++ frontend/src/pages/AdminPage/tabs/PhotoEditTab/components/desktop/ImagePreview/ImagePreview.styles.ts
